requirements.txt
transformers[torch]
fastapi
uvicorn
jinja2
google-generativeai
python-dotenv
pytest
loguru
typing_extensions